Run `forge fmt`

+++ src/interfaces/uniswap/IUniversalRouter.sol
@@ -7,7 +7,7 @@ import {IRewardsCollector} from './IRewardsCollector.sol';
 
 interface IUniversalRouter is IRewardsCollector, IERC721Receiver, IERC1155Receiver {
     /// @notice Thrown when a required command has failed
-    error ExecutionFailed(uint256 commandIndex, bytes message);
+    error ExecutionFailed(uint commandIndex, bytes message);
 
     /// @notice Thrown when attempting to send ETH directly to the contract
     error ETHNotAccepted();
@@ -22,5 +22,5 @@ interface IUniversalRouter is IRewardsCollector, IERC721Receiver, IERC1155Receiv
     /// @param commands A set of concatenated commands, each 1 byte in length
     /// @param inputs An array of byte strings containing abi encoded inputs for each command
     /// @param deadline The deadline by which the transaction must be executed
-    function execute(bytes calldata commands, bytes[] calldata inputs, uint256 deadline) external payable;
+    function execute(bytes calldata commands, bytes[] calldata inputs, uint deadline) external payable;
 }
